+++ src/context/AuthenticationProvider.tsx
@@ -1,16 +1,16 @@
-import React, { createContext, useState, useEffect, useContext } from "react";
+import React, { createContext, useState, useEffect } from "react";
 import { UserToken } from "@/models/AuthModels/userToken";
 
 const TOKEN_STORAGE_KEY = "token";
 
-interface AuthContextProps {
+export interface AuthContextProps {
   accessToken: string | null;
   user: UserToken | null;
   saveToken: (token: string) => void;
   clearToken: () => void;
 }
 
-const AuthContext = createContext<AuthContextProps | undefined>(undefined);
+export const AuthContext = createContext<AuthContextProps | undefined>(undefined);
 
 export const AuthProvider: React.FC<{ children: React.ReactNode }> = ({ children }) => {
   const [accessToken, setAccessToken] = useState<string | null>(null);
@@ -45,12 +45,4 @@ export const AuthProvider: React.FC<{ children: React.ReactNode }> = ({ children
       {children}
     </AuthContext.Provider>
   );
-};
-
-export const useAuth = (): AuthContextProps => {
-  const context = useContext(AuthContext);
-  if (!context) {
-    throw new Error("useAuth must be used within an AuthProvider");
-  }
-  return context;
 };
