+++ src/context/useAuth.ts
@@ -1,0 +1,10 @@
+import { useContext } from "react";
+import { AuthContext, AuthContextProps } from "./AuthenticationProvider";
+
+export const useAuth = (): AuthContextProps => {
+  const context = useContext(AuthContext);
+  if (!context) {
+    throw new Error("useAuth must be used within an AuthProvider");
+  }
+  return context;
+};
